+++ cosmohub/api/rest/downloads.py
@@ -15,7 +15,7 @@ from cosmohub.api import db, api_rest
 
 from ..db import model
 from ..db.session import transactional_session
-from ..security import auth
+from ..security import auth_required, PRIV_USER, PRIV_QUERY_DOWNLOAD
 from ..io.hdfs import HDFSPathReader
 
 def create_content_range(range_header, length):
@@ -111,7 +111,7 @@ class BaseDownload(object):
         return response
 
 class DatasetReadmeDownload(BaseDownload, Resource):
-    decorators = [auth.login_required]
+    decorators = [auth_required(PRIV_USER)]
 
     @staticmethod
     def _get_path(item):
@@ -145,7 +145,7 @@ class DatasetReadmeDownload(BaseDownload, Resource):
 api_rest.add_resource(DatasetReadmeDownload, '/downloads/datasets/<int:id_>/readme')
 
 class FileResource(Resource):
-    decorators = [auth.login_required]
+    decorators = [auth_required(PRIV_USER)]
 
     def get(self, id_):
         with transactional_session(db.session, read_only=True) as session:
@@ -197,7 +197,7 @@ class FileContentsDownload(BaseDownload, FileResource):
 api_rest.add_resource(FileContentsDownload, '/downloads/files/<int:id_>/contents')
 
 class QueryDownload(BaseDownload, Resource):
-    decorators = [auth.login_required]
+    decorators = [auth_required(PRIV_USER | PRIV_QUERY_DOWNLOAD)]
 
     def _headers(self, path):
         headers = super(QueryDownload, self)._headers(path)