+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Return total number of rows of a catalog. (Pau Tallada)
+- Token authentication and authorization. (Pau Tallada)
 
 
 ## [1.0.0] - 2016-08-19